Potential issue in VirtualDisplay due to NUM_BUFFER_SLOTS increased to 64
BZ: 128174

Change http://android.intel.com:8080/87673 increased the value of the
constant NUM_BUFFER_SLOTS from 32 to 64. The VirtualDisplaySurface again
use uint32 as a 'mask' to track buffer usage. Change it to uint64 to avoid
potential crashes.

Category: aosp improvement
Domain: Framework-GUI
Origin: internal
Upstream-Candidate: no, proprietary

+++ services/surfaceflinger/DisplayHardware/VirtualDisplaySurface.cpp
@@ -273,7 +273,7 @@ status_t VirtualDisplaySurface::dequeueBuffer(Source source,
     int pslot = mapSource2ProducerSlot(source, *sslot);
     VDS_LOGV("dequeueBuffer(%s): sslot=%d pslot=%d result=%d",
             dbgSourceStr(source), *sslot, pslot, result);
-    uint32_t sourceBit = static_cast<uint32_t>(source) << pslot;
+    uint64_t sourceBit = static_cast<uint64_t>(source) << pslot;
 
     if ((mProducerSlotSource & (1u << pslot)) != sourceBit) {
         // This slot was previously dequeued from the other source; must
